V17.19
Adaptation calcul puissance pour le triphasé UxIx3

+++ Source_MQTT.ino
@@ -40,7 +40,7 @@ void UpdatePmqtt() {
     }
     PVAS_M_inst = 0;
     EAIfloat += -Pw / 6000.0;
-    Energie_M_Injectee = int(EAIfloat);
+    Energie_M_Injectee = long(EAIfloat);
   }
 
   filtre_puissance();
